#35-refactor(stats): 가독성과 유지보수성을 위해 코드를 기능별로 파일 분리
