+++ auth/api.py
@@ -4,7 +4,7 @@ MVP demo ver 0.0.4
 auth/api.py
 
 역할: DRF REST API
-목적: views.py 파일 내의 복잡성을 줄이고, 인증 관련 로직을 별도의 파일로 분리하기 위해 만든 파일
+목적: participants_view.py 파일 내의 복잡성을 줄이고, 인증 관련 로직을 별도의 파일로 분리하기 위해 만든 파일
 기능: JWT 인증을 사용한 로그인, 로그아웃, 토큰 갱신, 액세스 토큰 만료시 리프레시 토큰과 함께 반환하도록 설정
 '''
