+++ participants/views/__init__.py
@@ -1,0 +1,2 @@
+from .participants_view import ParticipantViewSet
+from .statistics_view import StatisticsViewSet
